src/hekate/requirements.txt
redis==5.0.1
pyyaml==6.0.1
prometheus-client==0.19.0
pytest==7.4.3